Making Lua files translatable, and sorting strings by filename.
